requirements.txt
fastapi>=0.115,<1
uvicorn>=0.32,<1
grpcio>=1.67.1,<2
grpcio-tools>=1.67.1,<2
websockets>=15.0.1,<16
python-multipart>=0.0.20,<1
nvidia-riva-client>=2.25.1,<3
miniaudio>=1.61,<2
